Forgot to divide this up into commits.
I just found this in my working director, apparently I had forgotten to
divide this up into commits while I was working on it. There are
probably only 2 or 3 commits here and they won't overlap in files, so
the changes should be distinguishable if need be.

This is why we have tests.